Django/django-set/pyproject.toml
[tool.poetry]
name = "django-set"
version = "0.1.0"
description = ""
authors = ["KIM SHINEDAY <kimshineday99@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.3"
djangorestframework = "^3.16.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
